move RocksDBOptionsFeature to where it belongs
