+++ arangod/RocksDBEngine/RocksDBTransactionCollection.cpp
@@ -30,9 +30,9 @@
 #include "Random/RandomGenerator.h"
 #include "RocksDBEngine/RocksDBIndex.h"
 #include "RocksDBEngine/RocksDBMetaCollection.h"
+#include "RocksDBEngine/RocksDBOptionFeature.h"
 #include "RocksDBEngine/RocksDBSettingsManager.h"
 #include "Statistics/ServerStatistics.h"
-#include "StorageEngine/RocksDBOptionFeature.h"
 #include "StorageEngine/TransactionState.h"
 #include "Transaction/Hints.h"
 #include "Transaction/Methods.h"
